자료구조 분리
